fixed some bugs
for some reason, computer.pushSignal() didn't like the raw table so lets just give it a serialized one since that works more reliably.

also it would seem that my laptop is set to use tabs for indentation, not spaces.

also floor'd the speed so it's not got lots of decimal places after it. throttle and break have been left alone because their range is only 0-1