Well, we're getting somewhere!
detection.lua: changed how the data is sent over the network. Now only the augment address is sent over and the stock ID and data are now two seperate entries rather than the Id being the key. (makes it easier to extract it at the other end)

controller file: not done yet, but getting there. uses FIFO queues to check stock in and out of the block and will process the needed messages to the signals for them to work (that'll be in the next commit)